+++ src/app/home/components/MyLocationWeather/MyLocationWeather.tsx
@@ -32,22 +32,6 @@ const WEATHER_ICON = {
   없음: null,
 };
 
-const findDust = (pm10: number | null) => {
-  if (!pm10) {
-    return;
-  }
-
-  if (pm10 >= 0 && pm10 <= 30) {
-    return "좋음";
-  } else if (pm10 >= 31 && pm10 <= 80) {
-    return "보통";
-  } else if (pm10 >= 81 && pm10 <= 150) {
-    return "나쁨";
-  } else if (pm10 > 150) {
-    return "매우 나쁨";
-  }
-};
-
 interface LocationType {
   lat: number | null;
   lng: number | null;
