+++ src/app/home/components/MyLocationWeather/utils/index.ts
@@ -1,0 +1,1 @@
+export { default as findDist } from "./findDust";
